moved escape and unescape to utils module
